Move NonlinearityLayer from layers/dense to layers/special
