+++ lasagne/layers/special.py
@@ -1,13 +1,41 @@
 import theano
 import theano.tensor as T
 
-from .base import MergeLayer
+from .. import nonlinearities
+from .base import Layer, MergeLayer
 
 
 __all__ = [
+    "NonlinearityLayer",
     "InverseLayer",
     "TransformerLayer",
 ]
+
+
+class NonlinearityLayer(Layer):
+    """
+    lasagne.layers.NonlinearityLayer(incoming,
+    nonlinearity=lasagne.nonlinearities.rectify, **kwargs)
+
+    A layer that just applies a nonlinearity.
+
+    Parameters
+    ----------
+    incoming : a :class:`Layer` instance or a tuple
+        The layer feeding into this layer, or the expected input shape
+
+    nonlinearity : callable or None
+        The nonlinearity that is applied to the layer activations. If None
+        is provided, the layer will be linear.
+    """
+    def __init__(self, incoming, nonlinearity=nonlinearities.rectify,
+                 **kwargs):
+        super(NonlinearityLayer, self).__init__(incoming, **kwargs)
+        self.nonlinearity = (nonlinearities.identity if nonlinearity is None
+                             else nonlinearity)
+
+    def get_output_for(self, input, **kwargs):
+        return self.nonlinearity(input)
 
 
 class InverseLayer(MergeLayer):
